requirements/test.txt
-r base.txt

# factory boy
Faker==2.0.4
factory-boy==2.12.0
python-dateutil==2.8.1
text-unidecode==1.3
entrypoints==0.3

# flake8
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1

# isort
flake8-isort==2.7.0
isort==4.3.21
testfixtures==6.10.3

# pytest
attrs==19.3.0
coverage==4.5.4
more-itertools==7.2.0
packaging==19.2
parameterized==0.7.1
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
pytest-cov==2.8.1
pytest-django==3.7.0
